requirements.txt
python-telegram-bot==20.3  # Нова версія бібліотеки (синтаксис майже той самий, але з покращеннями)  
yt-dlp==2023.11.16         # Актуальний аналог youtube-dl (швидше і підтримує YouTube)  
sqlalchemy==2.0.23         # Сучасна робота з БД (ORM)  
python-dotenv==1.0.0       # Для зручного зчитування конфігурації з .env  
loguru==0.7.2              # Просте і потужне логування  